requirements.txt
python-telegram-bot>=12.0.0
SQLAlchemy>=1.1.0
requests>=2.0.0
beautifulsoup4>=4.0.0
